% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codelist_panel}
\alias{codelist_panel}
\title{List of Names and ISO Code for Countries}
\format{
\subsection{\code{codelist_panel}}{

A data frame with 28941 rows and 55 columns:
\describe{
\item{country.name.en}{Country name in English}
\item{iso2c, iso3c}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_countries_and_territories_by_land_and_maritime_borders}
}
\usage{
codelist_panel
}
\description{
List of Names and ISO Code for Countries
}
\keyword{datasets}
