% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_region_names}
\alias{country_region_names}
\title{Datasets of country and region names.}
\format{
\code{country_region_names}: A data frame with 644 rows and 3 variables
}
\usage{
country_region_names
}
\description{
Datasets of country and region names.
}
\author{
Yingjie Li \email{yingjieli.edu@gmail.com}
}
\keyword{dataset}
