% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_region.R
\name{detect_region}
\alias{detect_region}
\title{Detect country or region names in text for further mapping}
\usage{
detect_region(x, col)
}
\arguments{
\item{x}{Data frame or a string}

\item{col}{Column name for text to be assessed}
}
\value{
Returns the tool text outputs.
}
\description{
Detect country or region names in text for further mapping.
}
\examples{
x <- c("This paper explores the method and results from an independent
evidence based assessment of Australia's progress towards the SDGs",
"Last year alone, the United States experienced 14 separate billion-dollar
 disasters related to climate change")
col <- data.frame(x)
regions <- detect_region(x, col)

}
