% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_SDG_search_terms.R
\name{func_AND_vector}
\alias{func_AND_vector}
\title{\strong{Last update on}:  3/31/2022}
\usage{
func_AND_vector(v)
}
\arguments{
\item{v}{a vector of characters}
}
\value{
A character
}
\description{
\strong{New changes}:
}
\details{
Compare to the earlier version, we made the following changes
\enumerate{
\item Instead of combining multiple term lists by \code{OR} for one particular target,
it is more intuitive and accurate to add each alternative term list to the search
term table or database directly.
\item Added \verb{Look around} function to more accurately match SDG targets.
}

Use \code{AND} to Concatenate a Vector of Terms
}
\examples{
words <- c('apple', 'bean', 'food')
func_AND_vector(v= words)
}
