% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_SDG_search_terms.R
\name{func_OR_vector}
\alias{func_OR_vector}
\title{Use \code{OR} to Concatenate a Vector of Terms}
\usage{
func_OR_vector(v)
}
\arguments{
\item{v}{a vector of characters}
}
\value{
A character
}
\description{
Use \code{OR} to Concatenate a Vector of Terms
}
\examples{
words <- c('apple', 'bean', 'food')
func_AND_vector(v= words)
}
