% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_of_un_goals_targets}
\alias{list_of_un_goals_targets}
\title{The Names, ID, and Descriptions of all the 17 SDGs and 169 Targets}
\format{
\subsection{\code{list_of_un_goals_targets}}{

A data frame with 169 rows and 3 columns:
\describe{
\item{GoalID}{The ID of each SDG}
\item{GoalName}{The name of each SDG}
\item{target_id_un}{The name of each Target}
\item{target_desc_un}{The description for each Target}
}
}
}
\source{
\url{https://unstats.un.org/sdgs/indicators/indicators-list/}
}
\usage{
list_of_un_goals_targets
}
\description{
The Names, ID, and Descriptions of all the 17 SDGs and 169 Targets
}
\keyword{datasets}
