% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdg_bar.R
\name{plot_sdg_bar}
\alias{plot_sdg_bar}
\title{SDG bar plot}
\usage{
plot_sdg_bar(data, sdg = "sdg", value = "value", quiet = FALSE)
}
\arguments{
\item{data}{Data frame as the input}

\item{sdg}{Vector with SDG code to be visualized.}

\item{value}{The value, e.g., number of SDGs, to be show in the thematic map}

\item{quiet}{Logical. Suppress info message}
}
\value{
Returns the tool text outputs.
}
\description{
SDG bar plot
}
\examples{
data("sdgstat")
plot_sdg_bar(sdgstat, sdg = "SDG", value = "Value")

}
