% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdg_map.R
\name{plot_sdg_map}
\alias{plot_sdg_map}
\title{SDG Map Plot}
\usage{
plot_sdg_map(data, sdg = sdg, value = value,
                    country = country, by_sdg = TRUE)
}
\arguments{
\item{data}{Data frame as the input}

\item{sdg}{Vector with SDG code to be visualized.}

\item{value}{The value, e.g., number of SDGs, to be show in the thematic map}

\item{country}{Country that are associated with the SDGs.}

\item{by_sdg}{If mapping by SDG, TRUE or FALSE.}
}
\value{
Returns the tool text outputs.
}
\description{
SDG map plot
}
\examples{
data("sdgstat")
plot_sdg_map(sdgstat, sdg = "SDG", value = "Value",
             country = "Country", by_sdg = FALSE)

}
