% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_color.R
\name{sdg_color}
\alias{sdg_color}
\title{Color scheme for the 17 SDGs}
\usage{
sdg_color(x)
}
\arguments{
\item{x}{A number, which indicates the SDG ID}
}
\value{
HTML color code of a specified SDG
}
\description{
Color scheme for the 17 SDGs
}
\examples{
sdg_color(1)
sdg_color(x = 1:17)
}
