% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_icon.R
\name{sdg_icon}
\alias{sdg_icon}
\title{Icons for SDGs}
\usage{
sdg_icon(x, res = 200)
}
\arguments{
\item{x}{Numeric code for each SDG, ranging from 1 to 17}

\item{res}{Resolution of SDG icon. Default: \code{res = 200} indicates
resizing proportionally to 200px}
}
\description{
The \code{sdg_icon}  function provides the specific icon for each SDG
}
\examples{
sdg_icon(x = 17, res = 300)

}
