% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdg_icons}
\alias{sdg_icons}
\title{List SDG Icons}
\format{
\code{sdg_icons}: External pointer of class "magick-image"
}
\source{
\url{https://www.un.org/sustainabledevelopment/wp-content/uploads/2019/01/SDG_Guidelines_AUG_2019_Final.pdf}
}
\usage{
sdg_icons
}
\description{
List SDG Icons
}
\keyword{datasets}
