% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdgstat}
\alias{sdgstat}
\title{Datasets of SDG statistics.}
\format{
\code{sdgstat}: A data frame with 62 rows and 4 variables
}
\usage{
sdgstat
}
\description{
Datasets of SDG statistics.
}
\author{
Yingjie Li \email{yingjieli.edu@gmail.com}
}
\keyword{dataset}
