% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp}
\alias{shp}
\title{Datasets of shapefiles.}
\format{
\code{shp}: A data frame with 241 rows and 6 variables
}
\usage{
shp
}
\description{
Datasets of shapefiles..
}
\author{
Yingjie Li \email{yingjieli.edu@gmail.com}
}
\keyword{dataset}
