% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SDG_keys}
\alias{SDG_keys}
\title{Datasets of SDG keys.}
\format{
\code{SDG_keys}: A data frame with 557 rows and 3 variables
}
\usage{
SDG_keys
}
\description{
Datasets of SDG keys.
}
\author{
Yingjie Li \email{yingjieli.edu@gmail.com}
}
\keyword{dataset}
